Start work on UI and some additional refactoring
Begin implementing UI.js methods

Change Game.setup to Game.start

Add Util javascript to handle changing a board to a string. Putting this
in a util class keeps the toString method open on my board class.

Move valid number of players to be in the Game class. It makes no sense
for UI to decide how many players there should be.