Add human player and board to string
Human player asks for a move in a human readable way.

Board to string is a first go at printing a string representation of a
board. Should work with any size board.

Adding these two things allows me to start running player vs. player and
refactoring the game loop accordingly.

Change references to this to the local module.