Add end to the Game and refactor Game accordingly
This commit brings the game to a fully playable state.

Add an end method that is always called when the evaluator says the game
is over. This end method prints either a winning message with the
winning piece or a tie game message.

Change the name of the UI method called to be more clear.

Add an overall Game describe in the GameSpec to maintain proper scoping
and follow the normal describe conventions.